refactor: move signatures suites to vc module
