--- packages/core/src/crypto/signature-suites/bbs/types/GetProofsResult.ts
@@ -1,28 +1,0 @@
-/*
- * Copyright 2020 - MATTR Limited
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *     http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import type { JsonArray, JsonObject } from '../../../../types'
-
-/**
- * Result for getting proofs from a JSON-LD document
- */
-export interface GetProofsResult {
-  /**
-   * The JSON-LD document with the linked data proofs removed.
-   */
-  document: JsonObject
-  /**
-   * The list of proofs that matched the requested type.
-   */
-  proofs: JsonArray
-}
