+++ packages/core/src/modules/vc/SignatureSuiteRegistry.ts
@@ -1,8 +1,9 @@
-import { suites } from '../../../types/jsonld-signatures'
 import { KeyType } from '../../crypto'
-import { Ed25519Signature2018 } from '../../crypto/signature-suites'
-import { BbsBlsSignature2020, BbsBlsSignatureProof2020 } from '../../crypto/signature-suites/bbs'
 import { AriesFrameworkError } from '../../error'
+
+import { suites } from './libraries/jsonld-signatures'
+import { Ed25519Signature2018 } from './signature-suites'
+import { BbsBlsSignature2020, BbsBlsSignatureProof2020 } from './signature-suites/bbs'
 
 const LinkedDataSignature = suites.LinkedDataSignature
