+++ packages/core/src/modules/vc/signature-suites/bbs/BbsBlsSignatureProof2020.ts
@@ -11,8 +11,9 @@
  * limitations under the License.
  */
 
-import type { JsonObject } from '../../../types'
-import type { DocumentLoader, Proof } from '../../../utils'
+import type { JsonObject } from '../../../../types'
+import type { Proof } from '../../jsonldUtil'
+import type { DocumentLoader } from '../../libraries/jsonld'
 import type { DeriveProofOptions, VerifyProofOptions, CreateVerifyDataOptions, CanonizeOptions } from './types'
 import type { VerifyProofResult } from './types/VerifyProofResult'
 
@@ -20,11 +21,11 @@ import { blsCreateProof, blsVerifyProof } from '@mattrglobal/bbs-signatures'
 import { Bls12381G2KeyPair } from '@mattrglobal/bls12381-key-pair'
 import { randomBytes } from '@stablelib/random'
 
-import jsonld from '../../../../types/jsonld'
-import { suites } from '../../../../types/jsonld-signatures'
-import { AriesFrameworkError } from '../../../error'
-import { SECURITY_CONTEXT_URL } from '../../../modules/vc/constants'
-import { TypedArrayEncoder } from '../../../utils'
+import { AriesFrameworkError } from '../../../../error'
+import { TypedArrayEncoder } from '../../../../utils'
+import { SECURITY_CONTEXT_URL } from '../../constants'
+import jsonld from '../../libraries/jsonld'
+import { suites } from '../../libraries/jsonld-signatures'
 
 import { BbsBlsSignature2020 } from './BbsBlsSignature2020'
