+++ packages/core/src/modules/vc/signature-suites/bbs/index.ts
