+++ packages/core/src/modules/vc/signature-suites/bbs/types/DidDocumentPublicKey.ts
