+++ packages/core/src/modules/vc/signature-suites/bbs/types/KeyPairSigner.ts
