+++ packages/core/src/modules/vc/signature-suites/bbs/types/KeyPairVerifier.ts
