+++ packages/core/src/modules/vc/signature-suites/ed25519/constants.ts
