+++ packages/core/src/utils/index.ts
@@ -9,4 +9,5 @@ export * from './regex'
 export * from './indyProofRequest'
 export * from './VarintEncoder'
 export * from './Hasher'
-export * from './jsonld'
+export * from './validators'
+export * from './type'
